+++ odoo/local-src/specific_stock/__manifest__.py
@@ -1,4 +1,3 @@
-
 # -*- coding: utf-8 -*-
 # Copyright 2016 Julien Coux (Camptocamp)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
@@ -10,25 +9,38 @@
     'license': 'AGPL-3',
     'category': 'Warehouse',
     'depends': [
+        'delivery_rounds',
         'product',
         'product_expiry',
-        'stock',
-        'stock_production_lot_expired_dates',
-        'stock_reception_priority',
-        'stock_receive_lot',
+        'purchase',
+        'sale',
+        'specific_helpdesk',
         'specific_purchase',
+        'stock',
         'stock_available_immediately',
+        'stock_product_bin',
+        'stock_production_lot_expired_dates',
+        'stock_receive_lot',
+        'stock_reception_priority',
     ],
     'website': 'http://www.camptocamp.com',
     'data': [
         # Views
         'views/product_category.xml',
         'views/product_template.xml',
+        'views/res_partner.xml',
+        'views/stock_backorder_reason.xml',
         'views/stock_location.xml',
         'views/stock_production_lot.xml',
         'views/stock_config_settings.xml',
         'views/stock_quant_package.xml',
         'wizards/stock_receive_lot.xml',
+
+        # Wizards
+        'wizards/stock_backorder_choice.xml',
+
+        # Security
+        'security/ir.model.access.csv',
 
         # Data
         'data/ir_cron.xml',