+++ odoo/local-src/specific_stock/tests/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
+from . import test_lot_checksum
+from . import test_picking_backorder
 from . import test_picking_transfer
 from . import test_product_category_expired_times
-from . import test_lot_checksum
 from . import test_stock_inventory